Requried changes are made
