--- src/modules/variables.js
@@ -1,3 +1,0 @@
-export const inputTask = document.getElementById('input-task');
-export const addBtn = document.getElementById('add-btn');
-export const taskList = document.getElementById('task-list');
